Rename namespace to just Fluent
